Remove Helpers feature from skit [skip ci]
The Helpers/ module (compile-once-link-many for shared codegen
functions) had no real consumer and added ~250 LOC, a cache layer,
3 CLI flags, and a `swiftc` step. Directory mode (`skit run InputDir/
-o OutDir/`) handles the N-inputs case without it. If shared codegen
becomes a real need, a simpler prelude-inline approach can be added
later.

Deletes: CompiledHelpers.swift, HelpersOptions.swift, the helpers
discovery/build/cache pipeline in Helpers.swift, the
--helpers/--no-helpers CLI flags, the helpers param threaded through
Runner.swift and OutputCache.outputCacheKey, and the Docs/skit.md
+ Sources/skit/README.md sections about helpers.

Renames Helpers.swift -> Toolchain.swift, keeping the surviving
toolchain/cache utilities (dylibFilename, captureSwiftVersion,
libStamp, syntaxKitCacheRoot) that OutputCache and FileManager
+IsLibDir still depend on.

Verified end-to-end: single-file mode renders correctly; directory
mode renders 3/3 in 0.77s cold, 0.28s warm cache, 0.65s --no-cache;
--help confirms helpers flags are gone.
