--- Sources/skit/HelpersOptions.swift
@@ -1,44 +1,0 @@
-//
-//  HelpersOptions.swift
-//  SyntaxKit
-//
-//  Created by Leo Dion.
-//  Copyright © 2026 BrightDigit.
-//
-//  Permission is hereby granted, free of charge, to any person
-//  obtaining a copy of this software and associated documentation
-//  files (the “Software”), to deal in the Software without
-//  restriction, including without limitation the rights to use,
-//  copy, modify, merge, publish, distribute, sublicense, and/or
-//  sell copies of the Software, and to permit persons to whom the
-//  Software is furnished to do so, subject to the following
-//  conditions:
-//
-//  The above copyright notice and this permission notice shall be
-//  included in all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
-//  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-//  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-//  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
-//  HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-//  WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-//  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-//  OTHER DEALINGS IN THE SOFTWARE.
-//
-
-#if canImport(Subprocess)
-
-  /// How `skit run` should treat a `Helpers/` directory for this invocation.
-  internal enum HelpersOptions {
-    /// Walk up from the input looking for `Helpers/`. The default.
-    case auto
-    /// Skip helpers discovery entirely (`--no-helpers`). The wrapped input
-    /// won't be able to `import SyntaxKitHelpers`.
-    case disabled
-    /// Use the directory at the given path (`--helpers <dir>`). Validated as
-    /// an existing directory in `CompiledHelpers.init`.
-    case explicit(String)
-  }
-
-#endif
